Multi-session work: warning cleanup, BP1045 marker resolution, and behavioral fixes
Consolidates uncommitted work across several sessions on the warning-cleanup branch:

- BP1045 (TODO/FIXME) marker campaign: informational rewords, dead-code
  removal, and resolution across libslic3r + imgui.
- Behavioral bug fixes (compiled clean; several runtime/debugger-verified):
  * LayerRegion.cpp: spiral_vase now accounts for raft layers when comparing
    against bottom_solid_layers (Layer::id() is raft-offset).
  * PrintConfig.cpp normalize_fdm: add missing create-if-absent flag on four
    opt<>() calls, fixing a null-deref crash when --spiral-vase is set via CLI
    without a full profile.
  * TreeSupport.cpp / OrcaTreeSupport3D.cpp: fix operator-precedence bug in the
    interface fill-angle (interface_angle was discarded; angle never alternated).
  * Model.cpp model_volume_list_changed: add mesh-content (shared_ptr identity)
    check so modifier-volume mesh swaps invalidate support/seam/brim reprocessing.
- Klipper time-estimator machine limits, GCodeProcessor test, raw-pointer
  ownership cleanups, and overflow-safe tree-support refactor (in-progress WIP).

+++ src/exif/exif.c
@@ -2765,7 +2765,7 @@ void setDefaultMPFSegmentHeader(MPF_HEADER* appHeader, const char* strId, uint16
  *  1: success
  *  0: error
  */
-static int readAppNSegmentHeader(FILE *fp, APP_HEADER* appHeader, size_t startOffset)
+static int readAppNSegmentHeader(FILE *fp, APP_HEADER* appHeader, long startOffset)
 {
     // read the APP1 header
     if (fseek(fp, startOffset, SEEK_SET) != 0 ||
@@ -2799,7 +2799,7 @@ static int readAppNSegmentHeader(FILE *fp, APP_HEADER* appHeader, size_t startOf
 *  1: success
 *  0: error
 */
-static int readMPFSegmentHeader(FILE *fp, MPF_HEADER* appHeader, size_t startOffset)
+static int readMPFSegmentHeader(FILE *fp, MPF_HEADER* appHeader, long startOffset)
 {
     // read the MPF header
     if (fseek(fp, startOffset, SEEK_SET) != 0 ||
